move page components.
