+++ client/src/components/pages/app/Footer.tsx
