+++ client/src/components/pages/app/page.tsx
@@ -2,11 +2,9 @@ import * as _ from 'lodash';
 import * as React from 'react';
 import { Redirect, Route, Switch } from 'react-router';
 import { HashRouter } from "react-router-dom";
-import Contact from '../../../Contact';
-import Footer from '../../../Footer';
-import Home from '../../../Home';
 import { IImageEntity } from '../../../model';
-import Profile from '../../../Profile';
+import { ContactPage, HomePage, ProfilePage } from '../../pages';
+import Footer from './Footer';
 import Header from './Header';
 
 interface IProps {
@@ -26,9 +24,9 @@ export class AppRootPage extends React.Component<IProps, {}> {
       <div className="container container-fluid">
         <Header image={headerImage} />
         <Switch>
-          <Route exact={true} path="/home" component={Home} />
-          <Route exact={true} path="/contact" component={Contact} />
-          <Route exact={true} path="/profile" component={Profile} />
+          <Route exact={true} path="/home" component={HomePage} />
+          <Route exact={true} path="/contact" component={ContactPage} />
+          <Route exact={true} path="/profile" component={ProfilePage} />
           <Redirect to='/home'/>
         </Switch>
         <Footer />
