+++ client/src/components/pages/contact/index.tsx
@@ -1,0 +1,1 @@
+export * from './page';
