+++ client/src/components/pages/home/index.tsx
@@ -1,0 +1,1 @@
+export * from './page';
