+++ client/src/components/pages/index.tsx
@@ -1,1 +1,4 @@
 export * from './app';
+export * from './home';
+export * from './profile';
+export * from './contact';
